Refactor selection code into reusable method.
